Move particle to _core
