+++ tests/test_diffusion.py
@@ -9,7 +9,7 @@ from parcels import Field, VectorField
 from parcels import FieldSet
 from parcels.interpolators import XLinear
 from parcels.kernels import AdvectionDiffusionEM, AdvectionDiffusionM1, DiffusionUniformKh
-from parcels.particle import Particle, Variable
+from parcels import Particle, Variable
 from parcels import ParticleSet
 from parcels import XGrid
 from tests.utils import create_fieldset_zeros_conversion
