+++ tests/test_particle.py
@@ -3,7 +3,13 @@ import pytest
 
 from parcels._core.utils.time import TimeInterval
 from parcels._datasets.structured.generic import TIME
-from parcels.particle import _SAME_AS_FIELDSET_TIME_INTERVAL, Particle, ParticleClass, Variable, create_particle_data
+from parcels._core.particle import (
+    _SAME_AS_FIELDSET_TIME_INTERVAL,
+    Particle,
+    ParticleClass,
+    Variable,
+    create_particle_data,
+)
 
 
 def test_variable_init():
